[Refactor] 대시보드 API 성능개선

+++ src/main/java/com/fcc/PureSync/context/member/controller/MypageController.java
@@ -3,7 +3,9 @@ package com.fcc.PureSync.context.member.controller;
 import com.fcc.PureSync.context.member.dto.MemberInfoDto;
 import com.fcc.PureSync.context.member.dto.MemberInfoUpdateDto;
 import com.fcc.PureSync.context.member.dto.MemberPasswordDto;
-import com.fcc.PureSync.dto.ResultDto;
+import com.fcc.PureSync.common.ResultDto;
+import com.fcc.PureSync.exception.CustomException;
+import com.fcc.PureSync.exception.CustomExceptionCode;
 import com.fcc.PureSync.jwt.CustomUserDetails;
 import com.fcc.PureSync.context.member.service.MypageService;
 import lombok.RequiredArgsConstructor;
@@ -23,6 +25,10 @@ public class MypageController {
     // 내 정보
     @GetMapping
     public ResultDto myInfo(@AuthenticationPrincipal CustomUserDetails customUserDetails) {
+        if (customUserDetails == null) {
+            throw new CustomException(CustomExceptionCode.INVALID_JWT);
+        }
+
         return mypageService.getMyInfo(customUserDetails.getMemSeq());
     }
 
@@ -31,6 +37,10 @@ public class MypageController {
     public ResultDto myPosts(
             @RequestParam(required = false, defaultValue = "0", value = "page") int pageNo,
             @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+        if (customUserDetails == null) {
+            throw new CustomException(CustomExceptionCode.INVALID_JWT);
+        }
+
         return mypageService.getPosts(pageNo, customUserDetails.getMemSeq());
     }
 
@@ -39,6 +49,9 @@ public class MypageController {
     public ResultDto myPostsLikes(
             @RequestParam(required = false, defaultValue = "0", value = "page") int pageNo,
             @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+        if (customUserDetails == null) {
+            throw new CustomException(CustomExceptionCode.INVALID_JWT);
+        }
         return mypageService.myPostsLikes(pageNo, customUserDetails.getMemSeq());
     }
 
@@ -49,25 +62,36 @@ public class MypageController {
             @RequestPart MemberInfoDto memberInfo,
             @RequestPart(value = "file", required = false) MultipartFile file
     ) throws IOException {
+        if (customUserDetails == null) {
+            throw new CustomException(CustomExceptionCode.INVALID_JWT);
+        }
         return mypageService.updateMemberInfo(customUserDetails.getMemSeq(), memberInfo, file);
     }
 
     // 내 신체정보 업데이트
     @PostMapping
     public ResultDto insertMemberBodyInfo(@AuthenticationPrincipal CustomUserDetails customUserDetails, @RequestBody MemberInfoUpdateDto dto) {
+        if (customUserDetails == null) {
+            throw new CustomException(CustomExceptionCode.INVALID_JWT);
+        }
         return mypageService.insertMemberBodyInfo(customUserDetails.getMemSeq(), dto);
     }
 
     // 내정보 업데이트 (기본정보 - 패스워드)
     @PutMapping("/password")
     public ResultDto updateNick(@AuthenticationPrincipal CustomUserDetails customUserDetails, @RequestBody MemberPasswordDto dto) {
+        if (customUserDetails == null) {
+            throw new CustomException(CustomExceptionCode.INVALID_JWT);
+        }
         return mypageService.updatePassword(customUserDetails.getMemSeq(), dto);
     }
 
     // 회원 탈퇴
     @DeleteMapping
-    public ResultDto deleteMember(
-            @AuthenticationPrincipal CustomUserDetails customUserDetails) {
+    public ResultDto deleteMember(@AuthenticationPrincipal CustomUserDetails customUserDetails) {
+        if (customUserDetails == null) {
+            throw new CustomException(CustomExceptionCode.INVALID_JWT);
+        }
         return mypageService.deleteMember(customUserDetails.getMemSeq());
     }
